# Improve PHP parsing performance