--- a/package.json
+++ b/package.json
@@ -70,7 +70,7 @@
     "node-sass": "^4.5.0",
     "pegjs": "^0.10.0",
     "pegjs-loader": "^0.5.1",
-    "phpegjs": "1.0.0-beta3",
+    "phpegjs": "1.0.0-beta5",
     "postcss-loader": "^1.3.3",
     "prismjs": "^1.6.0",
     "raw-loader": "^0.5.1",