--- a/phpcs.xml.dist
+++ b/phpcs.xml.dist
@@ -31,6 +31,9 @@
 	<rule ref="Squiz.Commenting.FileComment.Missing">
 		<exclude-pattern>phpunit/*</exclude-pattern>
 	</rule>
+	<rule ref="Squiz.Commenting.ClassComment.Missing">
+		<exclude-pattern>phpunit/*</exclude-pattern>
+	</rule>
 	<rule ref="Squiz.Commenting.ClassComment.SpacingAfter">
 		<exclude-pattern>phpunit/*</exclude-pattern>
 	</rule>