--- a/phpunit/class-performance-test.php
+++ b/phpunit/class-performance-test.php
@@ -17,19 +17,19 @@ if ( getenv( 'RUN_SLOW_TESTS' ) ) {
 				dirname( __FILE__ ) . '/fixtures/long-content.html'
 			);
 			$parser = new Gutenberg_PEG_Parser;
-			$result = $parser->parse( $html );
+			$parser->parse( $html );
 
 			$time = microtime( true ) - $start;
 			$mem = memory_get_usage() - $start_mem;
 
-			error_log( '' );
-			error_log( 'Memory used (KB) : ' . round( $mem / 1024 ) );
-			error_log( 'Time (seconds)   : ' . ( round( $time * 100 ) / 100 ) );
-
 			$this->assertLessThanOrEqual(
-				0.3,
+				0.3, // Seconds.
 				$time,
-				"Parsing 'phpunit/fixtures/long-content.html' took too long."
+				sprintf(
+					"Parsing 'phpunit/fixtures/long-content.html' took too long. Time: %s (seconds), memory used: %sKB.",
+					( round( $time * 100 ) / 100 ),
+					round( $mem / 1024 )
+				)
 			);
 		}
 	}