Fixes #1646 where the PHP parser introduced in #1152 caused a performance regression.

Previous parsing times displayed evidence of `O(n^2)` complexity - from https://github.com/WordPress/gutenberg/issues/1646#issuecomment-312515989:

```
 358 words: 0.06s
 921 words: 0.33s
1438 words: 0.77s
1793 words: 1.13s
```

Current parsing times (using a post not containing any WP blocks, generated the same way as the above example) are significantly improved, with `O(n)` complexity:

```
 500 words: 0.04s
1000 words: 0.08s
1500 words: 0.11s
2000 words: 0.13s
```

Read on for details about what changed.

---

Using [this excellent StackOverflow answer](https://stackoverflow.com/a/30604570), I set up [`xhprof`](http://php.net/manual/en/intro.xhprof.php) to do granular profiling of PHP code.  Here were the initial results when running `RUN_SLOW_TESTS=y phpunit -c phpunit-xhprof.xml --filter large_post`:

> <img src="https://user-images.githubusercontent.com/227022/27938692-f1cc94f6-62bf-11e7-955a-faa074e5b9be.png" width="450">

We're calling `mb_strlen` and `mb_substr` too many times (the post being tested is the Gutenberg demo content, which has around 10k characters), and these functions are too slow.

`mb_strlen` is really easy to fix, because it's always called against the full length of the input string.  Fixed in https://github.com/nylen/phpegjs/commit/9f6b0f8cbb30a03c1ba24aa17ab1c9c940dbb105.

`mb_substr` is a bit more difficult to get rid of, but it's important to realize what these multibyte string functions actually do.  They can't know the length of the string in advance, because UTF-8 characters are variable-length.  So, every substring operation needs to loop through the entire string until it reaches its starting position.  This is the source of the previous `O(n^2)` complexity.

This is fixed in https://github.com/nylen/phpegjs/commit/3bfd473ed9940b961db24628767c324c736b1e19 by splitting the parse input into an array of UTF-8 characters when a parsing run begins.  Direct array lookups are far faster than looping over a multibyte string, and [WordPress core already uses this approach in other places](https://core.trac.wordpress.org/browser/tags/4.8/src/wp-includes/compat.php?annotate=blame&marks=95-100#L82).  (Like core does in the code block below that one, we need to detect when PCRE Unicode support is not available, and use an alternative method.  This is a task for a separate PR.)

Final performance results indicate further room for improvement, and there are some more obvious things to try, such as eliminating the `input_substr` function for substrings that are known to be of length 1.  However, the major issues with the initial version are fixed and we can continue iterating later:

> <img src="https://user-images.githubusercontent.com/227022/27938889-6572def0-62c1-11e7-8b92-ab9406fd8920.png" width="450">