@westonruter 2017-07-07
@mtias I think the parsing performance problem should be fixed prior to the 0.4 release.