@nylen 2017-07-07
Yep, that's the plan :)

@westonruter I am pretty satisfied with test coverage on these changes; do you see any issues here?