@nylen 2017-07-07
Regarding the last couple of review comments here: I would like to build a process that runs these "slow tests" every hour or so against `master`, then reports the results somewhere. I plan for this to handle many of the cases described at https://github.com/WordPress/gutenberg/issues/939#issuecomment-305260781.

For performance tests, PHPunit may not be the best tool. Ideally the performance test mechanism should store past results, provide a history, and compare them against previous results _from the same machine_.

That's all for later - the important thing for now is to fix the performance degradation included in the 0.3.0 release.