@dmsnell 2017-07-10
Very good work here @nylen, and I'm sorry for jumping to a false conclusion earlier about the source of the performance issues.

Curious:
  - do we have any comparison to how fast the JS-based parser performs on the same documents on the same machine?
  - do you see any other major performance issues here?

In light of this I'm going to close our #1681 and try to address other parser-related tasks. Thanks for staying persistent on this.