@nylen 2017-07-10
> do we have any comparison to how fast the JS-based parser performs on the same documents on the same machine?

I just ran some more tests on parsing the fixture introduced in this PR.

```
JavaScript :  11ms
PHP 5.6    : 150ms
PHP 7.0    :  17ms
PHP 7.1    :  19ms
```

Fortunately PHP 7.x is our [official recommendation](https://wordpress.org/about/requirements/).

> do you see any other major performance issues here?

Nothing major really.

I am not a huge fan of the current approach of splitting the entire input string into an array at the beginning of the parse, but I don't see a good way around it, other than proving that we can safely use `substr` (sounds tricky and unlikely).

Fixing #1611 should provide further improvements (the `peg_regex_test` block at the bottom of the second performance graph).

We can get rid of a bunch more `input_substr` function calls (the red block at the bottom of the second performance graph). These calls aren't very complicated so I'm not sure how much it will actually help.