@dmsnell 2017-07-10
Ouch on 5.6 but those are good numbers otherwise. Keeping the PHP parse within a magnitude of the JS parse I think is a satisfactory target. Getting them within half the runtime of JS is good!